Fix error introduced in login.php
